Refactor the JAX dispatcher
