+++ aesara/link/jax/dispatch/__init__.py
@@ -1,0 +1,16 @@
+# isort: off
+from aesara.link.jax.dispatch.basic import jax_funcify, jax_typify
+
+# Load dispatch specializations
+import aesara.link.jax.dispatch.scalar
+import aesara.link.jax.dispatch.tensor_basic
+import aesara.link.jax.dispatch.subtensor
+import aesara.link.jax.dispatch.shape
+import aesara.link.jax.dispatch.extra_ops
+import aesara.link.jax.dispatch.nlinalg
+import aesara.link.jax.dispatch.slinalg
+import aesara.link.jax.dispatch.random
+import aesara.link.jax.dispatch.elemwise
+import aesara.link.jax.dispatch.scan
+
+# isort: on
